--- integration-tests/cxf-soap/src/test/resources/mappings/wsSecurity.json
@@ -1,59 +1,0 @@
-{
-  "request": {
-    "method": "POST",
-    "url": "/hellowss",
-    "headers": {
-      "Content-Type": {
-        "equalTo": "text/xml; charset=UTF-8"
-      },
-      "SOAPAction": {
-        "equalTo": "\"https://www.helloworld.com/Service/Hello\""
-      }
-    },
-    "bodyPatterns": [
-      {
-        "matchesXPath": "//wsse:Security",
-        "xPathNamespaces": {
-          "wsse": "http://docs.oasis-open.org/wss/2004/01/oasis-200401-wss-wssecurity-secext-1.0.xsd"
-        }
-      },
-      {
-        "matchesXPath": "//wsse:UsernameToken",
-        "xPathNamespaces": {
-          "wsse": "http://docs.oasis-open.org/wss/2004/01/oasis-200401-wss-wssecurity-secext-1.0.xsd"
-        }
-      },
-      {
-        "matchesXPath": "//wsse:Username",
-        "xPathNamespaces": {
-          "wsse": "http://docs.oasis-open.org/wss/2004/01/oasis-200401-wss-wssecurity-secext-1.0.xsd"
-        }
-      },
-      {
-        "matchesXPath": "//wsse:Password",
-        "xPathNamespaces": {
-          "wsse": "http://docs.oasis-open.org/wss/2004/01/oasis-200401-wss-wssecurity-secext-1.0.xsd"
-        }
-      },
-      {
-        "matchesXPath": "//wsse:Nonce",
-        "xPathNamespaces": {
-          "wsse": "http://docs.oasis-open.org/wss/2004/01/oasis-200401-wss-wssecurity-secext-1.0.xsd"
-        }
-      },
-      {
-        "matchesXPath": "//wsu:Created",
-        "xPathNamespaces": {
-          "wsu": "http://docs.oasis-open.org/wss/2004/01/oasis-200401-wss-wssecurity-utility-1.0.xsd"
-        }
-      }
-    ]
-  },
-  "response": {
-    "status": 200,
-    "bodyFileName": "wsSecurityResponse.xml",
-    "headers": {
-      "Content-Type": "text/xml; charset=UTF-8"
-    }
-  }
-}
