+++ integration-tests/pom.xml
@@ -78,7 +78,7 @@
         <module>crypto</module>
         <module>csimple</module>
         <module>csv</module>
-        <module>cxf-soap</module>
+        <module>cxf-soap-grouped</module>
         <module>dataformat</module>
         <module>dataformats-json</module>
         <module>datasonnet</module>
